[ingress][pytorch] Basic KernelBench to MLIR conversion
Basic as can be torch-mlir converter for the level1 and level2
KernelBench kernels. The `convert-kernel-bench-to-mlir.py` script does
the conversion and dumps the results in the `cache/level1` and
`cache/level2` folders.

Relies on pre-packaged mlir wheels and mlir-torch, as this PR considers
dealing with versioning and packaging an orthogonal matter to getting
ingress up and running.

About ~55 of the 200 kernels are filtered out as they either crash
torch-mlir or yield very big .mlir files. This ignore_list is meant to
be amended as these issues get addressed, e.g. by altering init_inputs
on a per kernel basis.

The conversion script sticks to outputting just linalg for now. As it
does this, it does do some basic post-processing of torch-mlir's output,
namely it runs the -linalg-specialize-generic-ops pass.