[Python] Basic dependency management with pyproject and uv (#6)
Sets us up to do a uv sync to set up the current Python environment with the correct dependencies. Note that this is just a facilitating mechanism for preparing the python environment; users are still free to set up the environment via any means they like.

By default the only dependency installed is the Python bindings (from llvm/eudsl) -- that is, enabling import mlir in Python. This is the common dependency for generating and running pipelines and for ingress. Optionally one can pass (a variant of) --extra ingress-torch to uv sync to also get the torch and torch-mlir dependencies -- that is, enabling import torch and import torch_mlir in Python. Per the included docs, it is recommended to first set up a fresh virtualenv, e.g. via uv venv.

The pyproject.toml comes with uv-specific sections so as to enable specifying from which "indices", i.e. Python package repositories, each specific package is expected to come from. This file pins particular nightlies/versions from these package repositories.