[CI] Linter and formatter (#24)
Adds 'ruff' as Python linter and code formatter, and 'pre-commit' for
easy rules application both locally and in CI.
New lint CI workflow is also added.

The formatting tools are added to the 'dev' dependency group.
They are installed by default with `uv sync` and can be run locally
with the following cmd:
  pre-commit run -a