[mlir-gen] Add mlir builders for llama3.1 and tests (#13)
Putting up this dirty draft for early feedback/questions. I'm putting
together some tests to run a e2e llama3.1 going through linalg on
tensors. The goal is to generate some nice linalg that would be
optimization friendly. At the moment, there are just functional blocks
and pieces that are just smoke-tested. These include naive
implementations for rotary embeddings, feed forward, rms, and a bunch of
other small snippets that are useful to implement the model. These are
already enough to put an attention block together. It'd be nice to test
it against the original implementation, but that'd require fairscale as
a dependency. For now I only added pytest and kept the pipeline as
simple as possible. I also reused the example with the schedule, so now
it is a part of every test.