[examples][ingress][torch] Basic KernelBench to MLIR conversion (#5)
Basic torch-mlir converter for the level1 and level2 KernelBench
kernels. The `convert-kernel-bench-to-mlir.py` script does the
conversion and dumps the results in `level1` and `level2` dirs inside
the `cache/ingress/KernelBench` folder. Conversion is done per kernel,
with multiple conversions being run in parallel.

56 of the 200 kernels are filtered out as they either crash torch-mlir
or yield very big .mlir files. This ignore_list is meant to be amended
as these issues get addressed, e.g. by altering init_inputs on a per
kernel basis.

The conversion script sticks to outputting just linalg for now. As it
does this, it does do some basic post-processing of torch-mlir's output,
namely it runs the -linalg-specialize-generic-ops pass.